lexicon: first draft of new portability schema
