+++ .changeset/wacky-carrots-clean.md
@@ -1,0 +1,5 @@
+---
+'@openfn/lexicon': major
+---
+
+Introduce a formal Portabilty schema
